Feat: #196 Mock에 정의한 타입 분리
